Move `ecosystem_versions` to `FileParser` class
The intent of `ecosystem_versions` is to report the range of versions
supported by the target repo, _not the version that Dependabot chose to
use_.

In order to do that, we need to parse the manifest files for any lines
indicating the range of supported versions.

Intuitively, this sounds like `FileParser`, not `FileFetcher`.

Historically though, the code has lived in `FileFetcher`, where it sorta
grew organically. So as we systematize this, it makes more sense to move
it to the `FileParser`.